Added online Multiplayer
Used PUN.
To run in MP, you need to begin the game from the MenuScene Scene. It currently spawns you in as the Player1 Prefab.
To debug it in offline (which is faster for devs), drag a Player1 prefab onto the Level1 scene and check its "For Offline Testing" box. You can now run it in offline. To suppress a (harmless) error about "Creating a Player before Instantiating the Room," go to the GameSetup gameObject in the Hierarchy and check the "Turn Off Online" box.
I followed this guide https://www.red-gate.com/simple-talk/dotnet/c-programming/creating-multiplayer-games-with-unity-and-pun/